+++ RELEASE_NOTES.md
@@ -21,6 +21,7 @@
 
 ### 🐛 Bug Fixes / 修复
 
+- accept generated config
 - wait for docker health
 - bootstrap openclaw config
 - enable tini subreaper mode